Add dockerized web-app and adjust DNS setup
Introduced a Dockerized web-application set up and adjusted the DNS configuration to accommodate the changes. This included addition of nginx serving the web-app, exposing it on port 9878, and linking it to the DNS server in the docker-compose configuration. The DNS setup was adjusted as part of this change, moving from workspace.com zone to a local-network.app basis and setting new reverse DNS settings. These changes were made to facilitate a segregated testing environment with linked services for our application.